neural/pyproject.toml
[project]
name = "neural"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "openai>=1.93.0",
]
